transformer_tools/text_augmentor/requirements.txt
fasttext==0.9.2
jsonschema==3.2.0
lxml==4.6.3
matplotlib==3.4.1
nest-asyncio==1.5.1
notebook==6.3.0
numpy
pandas==1.2.3
pytest==6.2.3
pyyaml
python-Levenshtein==0.12.0
spacy
tqdm
boto3==1.16.53
mlflow==1.13.1
mysqlclient==2.0.3
optuna==2.4.0
pyyaml==5.4.1
scikit-learn==0.24.0
mltb==0.7.dev3
torch==1.7.1
transformers==4.4.0
torch
#fastBPE
regex
requests
#sacremosesd
#subword_nmt
#fairseq
nlpaug